how fix bug
