+++ app/gui2/src/components/VisualizationContainer.vue
@@ -147,13 +147,15 @@ const resizeBottomRight = usePointer((pos, _, type) => {
                 "
               />
             </button>
-            <VisualizationSelector
-              v-if="isSelectorVisible"
-              :types="config.types"
-              :modelValue="config.currentType"
-              @hide="isSelectorVisible = false"
-              @update:modelValue="(isSelectorVisible = false), config.updateType($event)"
-            />
+            <Suspense>
+              <VisualizationSelector
+                v-if="isSelectorVisible"
+                :types="config.types"
+                :modelValue="config.currentType"
+                @hide="isSelectorVisible = false"
+                @update:modelValue="(isSelectorVisible = false), config.updateType($event)"
+              />
+            </Suspense>
           </div>
         </div>
         <div v-if="$slots.toolbar" class="toolbar">
