+++ app/gui2/src/components/VisualizationSelector.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import SvgIcon from '@/components/SvgIcon.vue'
-import { useVisualizationStore } from '@/stores/visualization'
 import { useAutoBlur } from '@/util/autoBlur'
 import { visIdentifierEquals, type VisualizationIdentifier } from 'shared/yjsModel'
 import { onMounted, ref } from 'vue'
@@ -11,7 +10,10 @@ const props = defineProps<{
 }>()
 const emit = defineEmits<{ hide: []; 'update:modelValue': [type: VisualizationIdentifier] }>()
 
-const visualizationStore = useVisualizationStore()
+// This dynamic import is required to break the circular import:
+// `VisualizationSelector.vue` -> `compilerMessaging.ts` -> `VisualizationContainer.vue` ->
+// `VisualizationSelector.vue`
+const visualizationStore = (await import('@/stores/visualization')).useVisualizationStore()
 
 const rootNode = ref<HTMLElement>()
 useAutoBlur(rootNode)
