+++ app/gui2/src/components/visualizations/ImageBase64Visualization.vue
@@ -10,9 +10,8 @@ interface Data {
 </script>
 
 <script setup lang="ts">
+import { VisualizationContainer } from '@/util/visualizationBuiltins'
 import { computed } from 'vue'
-
-import VisualizationContainer from '@/components/VisualizationContainer.vue'
 
 const props = defineProps<{ data: Data }>()
