+++ app/gui2/src/components/visualizations/TableVisualization.vue
@@ -60,7 +60,7 @@ interface UnknownTable {
 </script>
 
 <script setup lang="ts">
-import VisualizationContainer from '@/components/VisualizationContainer.vue'
+import { VisualizationContainer } from '@/util/visualizationBuiltins'
 import type {
   ColDef,
   ColumnResizedEvent,
