+++ app/gui2/src/providers/visualizationConfig.ts
@@ -1,9 +1,9 @@
-import type { URLString } from '@/stores/visualization/compilerMessaging'
+import { createContextStore } from '@/providers'
 import type { Icon } from '@/util/iconName'
+import type { URLString } from '@/util/urlString'
 import { Vec2 } from '@/util/vec2'
 import type { VisualizationIdentifier } from 'shared/yjsModel'
 import { reactive } from 'vue'
-import { createContextStore } from '.'
 
 export interface VisualizationConfig {
   /** Possible visualization types that can be switched to. */
