+++ app/gui2/src/util/urlString.ts
@@ -1,0 +1,1 @@
+export type URLString = `${string}:${string}`
