+++ app/gui2/src/util/visualizationBuiltins.ts
@@ -1,5 +1,4 @@
-import VisualizationContainer from '../components/VisualizationContainer.vue'
-
+import VisualizationContainer from '@/components/VisualizationContainer.vue'
 export { useVisualizationConfig } from '@/providers/visualizationConfig'
 export { defineKeybinds } from '@/util/shortcuts'
 export { VisualizationContainer }
