+++ app/gui2/ydoc-server/fileFormat.ts
@@ -14,8 +14,8 @@ export type VisualizationMetadata = z.infer<typeof visualizationMetadata>
 const visualizationMetadata = z
   .object({
     show: z.boolean().default(true),
-    project: visualizationProject.nullable(),
-    name: z.string().nullable(),
+    project: visualizationProject.optional(),
+    name: z.string().optional(),
   })
   .passthrough()
